swf: Reorganize Tag
Move `Tag` into separate module and alphabetize variants.
